fix(piko): stabilize motion wake and show incidents

+++ piko/power_logic.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 
 inline bool shouldIdle(bool idle, bool inApp, uint32_t now, uint32_t lastActivity, uint32_t timeout) {
-  return !idle && inApp && now - lastActivity >= timeout;
+  return !idle && inApp && static_cast<int32_t>(now - lastActivity) >= static_cast<int32_t>(timeout);
 }
 
 inline bool motionDetected(float x, float y, float z,
